Feature: Separate file to return express app instance
